requirements.txt
mineru
pytest
openai
python-dotenv
pymupdf
